requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
pyarrow>=12.0.0
pymongo>=4.6.0
redis>=4.5.0
pyyaml>=6.0.0
